+++ app/Http/Requests/BulkDestroyRequest.php
@@ -2,16 +2,16 @@
 
 namespace App\Http\Requests;
 
-use App\User;
+use App\Models\User;
 use Illuminate\Foundation\Http\FormRequest;
 
-class MassDeleteRequest extends FormRequest
+class BulkDestroyRequest extends FormRequest
 {
     public function rules()
     {
         return [
             'users'   => ['required','array'],
-            'users.*' => ['bail','required','integer','exists:App\User,id',
+            'users.*' => ['bail','required','integer','exists:App\Models\User,id',
             function ($attribute, $value, $fail) {
                 $user = User::find($value);
                 if (!$this->room->members()->find($value) or $this->room->owner->is($user)) {