+++ app/Http/Requests/BulkUpdateRequest.php
@@ -3,18 +3,18 @@
 namespace App\Http\Requests;
 
 use App\Enums\RoomUserRole;
-use App\User;
+use App\Models\User;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
 
-class MassUpdateRequest extends FormRequest
+class BulkUpdateRequest extends FormRequest
 {
     public function rules()
     {
         return [
             'role'    => ['required',Rule::in([RoomUserRole::USER,RoomUserRole::MODERATOR,RoomUserRole::CO_OWNER])],
             'users'   => ['required','array'],
-            'users.*' => ['required','integer','exists:App\User,id',
+            'users.*' => ['required','integer','exists:App\Models\User,id',
             function ($attribute, $value, $fail) {
                 $user = User::find($value);
                 if (!$this->room->members()->find($value) or $this->room->owner->is($user)) {